Add option to specify TTL/hop_limit for unicast peers and to check it
This commit allows specifying the TTL/hop limit to use when sending to
unicast peers, and allows a minimum and maximum TTL/hop limit to be
specified against each unicast peer so that the TTL/hop limit can be
checked to be within the specified range when an advert is received.

This does not alter the TTL/hop limit used when multicasting, which
remains at 255.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>