Add ~LST configuration option
Issue #1581 requested the addition of list blocks, so act like
sequence blocks, but with the values to be substituted being specified
as part of the list block.

The commit extends the requested feature by allowing more than one
variable to be set, for example:

  ~LST({IP, IP1}, {10,1},{20,4},{5,6},{12,8}) 192.168.${IP}.${IP1}

produces 4 lines, with IP=10 and IP1=1 for the first line, etc.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>