Rate limit respawning of processes
If a keepalived child process dies, unless the -R option has been
specified, the child process is respawned. If the child process has
a hard fault that causes it to terminate very quickly, the process
was continually respawned, filling logs and creating a high system
load.

This commit rate limits the respawning of the process. The first time
a child process dies, it is immediately respawned. The next time there
is a 1 second delay, and each subsequent time is doubled up to a limit
of 60 seconds.

If the child process runs for 1 minute or more, it is considered not to
have a hard fault, and the reload delay is reset to 0.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>