Add option to send DELETE to notify scripts etc on reload
If a VRRP instance is removed on a reload, keepalived used to send
FAULT to all the notiy mechanisms. Issue #1526 requested that it be
possible to differentiate betwen fault state and deleted.

Since previously FAULT was sent, that needs to remain the default,
but we now have an option, notify_deleted, that causes DELETED
notifies to be sent instead of FAULT when vrrp instances are deleted
on a reload.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>