check_dns: Fix checker timeout
The sands timeout for dns_send() was incorrectly calculated as the
sands timeout of the timer thread which had already expired when the
dns_connect_thread() was called. Once time_now was subtracted from the
expired sands, the timeout had underflowed and was a value close to
ULONG_MAX, effectively resulting in no timeout.

dns_connect_thread() now adds the checker connection timeout to time_now
so that the actual timeout is in the near future. dns_send() now also
ensures that when calculating the next timeout there is no underflow.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>