Add ${_HASH} and ${_BANG} configuration options
A definition one could such as
$COMMENT=#
doesn't work since the '#' (and anything following) will be removed as
a comment.

This commit adds ${_HASH} and ${_BANG} which are replaced with # and !
respectively, without a comment being started.

This makes it easy to create configuration files with a large number of
optional elements that can be enabled or disabled by simple definition
changes.

For example, one could have:
$TEST=${_HASH}

$TEST test config
which would disable "test config" since it would be commented out.

Changing the definition of $TEST to be
$TEST=
or even
$TEST=#${_HASH}
would enable "test config".

Note, $TEST=#${_HASH} resolves to $TEST= since the '#' character starts
a comment.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>