config: Allow 'include' statement to be on any line in config files
Issue #1670 identified that using an include statement in the middle
of a virtual_ipaddress block didn't work. It turned out that 'include'
could not be used in any block that was a list of multiple entries,
e.g. routes, rules, email addresses.

This commit restructures the way handling to opening and closing of
config files is handled, so that include statements can appear on any
line of the config files.