Following commit a157bde, log original message if log_all_failures not set
The log message '... failed on retry 3/3 (exited with status 1)' was not
ideal if log_all_failures is not set. The commit causes the message
'... failed after 3 retries (exited with status 1)' to be logged if
log_all_failures is not set, which is what was logged before commit
a157bde - "Fix misc_check log of failed retry check".

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>