ipvs: Fix configuring of sync daemon
The sync daemon used to be either master or backup, but an option
was added to allow it to run in both master and backup mode,
although this was still all handled in the vrrp process.

There was a race condition when using both master and backup because
the checker process was stopping the IPVS sync at startup, and the
VRRP process was starting it.

This commit changes the way it works, so that the VRRP process only
handles the sync daemon if it is set to track a VRRP instance, and
if the daemon runs in both master and backup mode, it is handled by
the checker process.

The commit also resolves issues when reloading the config if the
sync daemon configuration is changed.

Finally, lvs_timeouts are moved from being handled by the VRRP process
to the checker process, values are updated after a reload, and the
original values are read before any values are changed, and those values
are restored on exit, or when no longer configured after a reload.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>