Optimise MALLOC calls when sending SMTP messages
The code used to MALLOC a structure and then 4 further MALLOCs for
buffers which elements of the structure then pointed to. If the
structure and buffers were not deallocated before terminating, and
MALLOC MEM_CHECK debugging was enabled, if the shutdown emails had
not been able to be sent before final shutdown occurred, for each
message not sent 5 outstanding allocations were reported, and if
there were several problems, this made it look as though there were
a large number of problems.

This commit changes the code to use a single MALLOC that includes
the buffers, and sets pointers accordingly; it is also more efficient.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>