Add PTR_CAST ready for updating code to use it
Using PTR_CAST with default configuration should make no difference
to the code generated, but can suppress unnecessary compiler
warnings.

On architectures which require aligned access, e.g. 32 bit ARM (an
example being the Raspberry Pi), the compiler produces a very large number
of "cast increases required alignment of target type" warnings. PTR_CAST
will now, by default, cast via a void *, which suppresses the warning.

The test for whether casting via a void * is needed to stop the warnings
is done by configure, and so casting via a void * is only done if it is
needed to suppress the warnings. configure options --enable-cast-via-void
and --disable-cast-via-void can be used to override the default.

To generate the "cast increases required alignment of target type" warnings
on an architecture that does not require aligned access, use configure
options: --enable-strict-cast-align --disable-cast-via-void. This can be
useful because even though an architecture like x86_64 allows unaligned
access, it is less efficient than aligned access, and so it is desirable
to avoid it; it is also useful for seeing what a system that requires
aligned access would produce, since it is important that there is no
attempts at unaligned access.

The final option this commit introduces is configure option
--enable-cast-align-checks, which adds runtime code to check that a
pointer being cast is suitably aligned for the pointer it is being
cast to. This should only be used as a development option to ensure
that all casts are suitably aligned; testing with this found 22 instances
in the keepalived code of unaligned casts, which will be fixed in a
later patch.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>