Fix misaligned casts identified by using PTR_CAST
Using PTR_CAST with configure option --enable-cast-align-checks and
running the keepalived code found 22 instances of unaligned pointer
casts. Although these would not have caused a problem on i386/x86_64
systems, on ARMv5 systems it would have caused incorrect data to be
returned. On all systems misaligned access is inefficient, but on
architectures that generate a trap, they are very inefficient.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>