vrrp: Fix IPv6 neighbour advert unaligned access
Due to struct eth_header being 14 bytes long, attempting to follow
that in a single buffer with a struct ip6hdr, which requires 4 byte
alignment, causes unaligned access unless the struct eth_header is
forced to be 2 byte aligned but not 4 byte aligned.

This commit, instead of forcing an unnatural alignment, uses a
gather write (sendmsg()) so that each header structure is constructed
independently.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>