vrrp: Fix unaligned access for data from process events connector
The data received from the kernel process events connect comprises:
struct nlmsghdr		# 16 bytes, 4 byte alignment
struct cn_msg		# 20 bytes, 4 byte alignment
struct proc_event	# requires 8 byte alignment

This means that if the struct nlmsghdr is 8 byte aligned,
struct proc_event is misaligned. The kernel handles this by allocating
a buffer that is 8 byte aligned and skipping the first 4 bytes of the
buffer. This has the consequence that two chained netlink messages
cannot be received in a single read, since one or other of the proc_event
messages will be misaligned.

This commit uses a scatter read (recvmsg()) to read each part of the
netlink message into a struct that is appropriately aligned.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>