vrrp: ensure virtual routes are not lost during a reload
If a virtual route depended on a VIP and a reload changes the VIP,
the kernel would delete the route when the VIP is deleted, but not
reinstate the route if a new VIP allowed the route to work again.

For example:
  old VIP:	1.2.3.4/24 dev ens4
  route:	default via 1.2.3.1 dev ens4
  new VIP:	1.2.3.5/24 dev ens4

Deleting 1.2.3.4 causes the route to be deleted, but the kernel doesn't
notify keepalived. Therefore, when keepalived adds 1.2.3.5, it doesn't
know to re-add the route.

This commit makes keepalived attempt to re-add all virtual routes after
VIP changes on a reload.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>