vrrp: Be consistent about whether an IPv6 address is set or not
Issue #1716 identified the keepalived was setting an invalid IPv6
address after a reload if the address had previously been configured
and then removed from an interface. The cause of this was that
sometimes only the first 32 bits were checked or cleared, and sometimes
all 128 bits were checked or cleared.

This commit introduces IS_IP6_ADDR() and CLEAR_IP6_ADDR() macros to ensure
consistency; currently all 128 bits are not checked or cleared, but it
would be simple to change the macros to only consider the first 32 bits if
that were preferred. Checking/clearing all 128 bits is probably the right
way to do it, since ::/128 means unspecified address, and not ::/32.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>