vrrp: Ensure first advert sent before GARPs on transition to master
If we are transitioning to master and we are using a VMAC, for IPv4
we must ensure that any existing (lower priority) master removes its
VIPs before we send the gratuitous ARPs.

The only way to make the old master remove the VIPs is to send a
higher priority advert, so we need to do so before sending the GARPs.

The problem with sending GARPs before the VIPs have been removed, when
using a VMAC, is that we send a GARP request, but the old master, since
it has the VIP configured on the same MAC address, then sends a GARP reply,
which will cause any MAC caches in switches to be updated again to point
to the old master for the VMAC MAC address.

This solution isn't perfect because it depends on the old master removing
the VIPs before it receives the GARP request that we will send, but it is
certainly better than sending GARPs first, which will guarantee that the
old master will send GARP replies.

We probably need to use a firewall, or eBPF filter, to properly stop the
GARP replies being sent.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>