bfd: Add support for configuring timers to micro-second resolution
Previously, min_rx, min_tx and idle_tx could only be specified in
whole milli-seconds.

RFC7419 states: "Common Interval values to be: 3.3 msec, 10 msec, 20 msec,
50 msec, 100 msec, and 1 sec". In order to support 3.3 msec, we need to be
able to specify fractions of msec, so this commit adds support for
specifying the timers to 3 decimal places of msecs, in other words allowing
the resolution down to single micro-seconds, which is the resolution of
the timers in the BFD protocol.

In order to support this change, all logged values of the timers are now
logged in micro-seconds, instead of the previous milli-seconds.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>