check: Add UDP_CHECK payload data for sending and checking reply
Pedro Viton of Nokia, Madrid requested the ability to be able to
specify a payload for the UDP messages sent for a UDP_CHECK, in order
to avoid having to use a MISC_CHECK, with the associated overhead that
causes.

This commit extends that idea by allowing the maximum and minimum required
payload lengths for the returned payload to be specified, and also allows
parts of the returned payload to be checked to match configured values.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>