parser: report_config_error when failing opening a file
while using `reload_check_config`, it is useful to make sure to fail
when the config file or its includes are not possible to be open.

it will avoid a situation where a reload continues to happen while an
important config included cannot be open.

to avoid breaking existing configuration introduce `glob_strict` so the
user can explictly choose to fail.

Signed-off-by: William Dauchy <w.dauchy@criteo.com>