ipvs: Allow multiple MISC_CHECKs and use with FILE_CHECKs
Previously a MISC_CHECK set the weight of the real server based
on it's exit code. This meant that a real server could not have
more than one MISC_CHECK (not really a problem because a script
could combine all the MISC_CHECK scripts), but also MISC_CHECKs
and FILE_CHECKs could not be combined.

The old way that MISC_CHECK worked was the weight of the real server
was set to two less that the script exit code (0 set weight to 0,
and 1 put the read server info failed state).

If effect, the MISC_CHECK added:
   (exit_code - 2 - real server initial weight)
to the weight of the real server. For example, suppose the real
server had a weight of 10 and the script exited with status 20,
(20 - 2 - 10) was added to the real server weight, giving a new
weight of 18 (i.e. two less than the exit code of the script).

The MISC_CHECK now does precisely that, and adds:
  (exit_code - 2 - real server initial weight).
This allows multiple MISC_CHECKs to work together, and they can
also be combined with FILE_CHECKs.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>