core: Only read config files once when starting or reloading
Some users who have automatically updated configurations and frequent
reloads can experience a problem when using the reload config_check if
a config file changes between when the config check runs and is
successful, and then the main keepalived processes read the config
files.

This commit makes keepalived write the config to a memory based file
when the config is first read, and then all the other processes read
the memory based file, rather than re-reading the config files.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>