core: Add reload_file config option
The commit makes keepalived create the specified reload_file before
it starts reading the configuration files during a config reload
and deletes the file after it has finished reading the configuration
files.

If the reload_file is created by the user (or a user process) before
signalling keepalived to reload, the user (or user processs) can
monitor the existance of the file, and when it is removed can start
modifying the config files again.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>