core: Add configuration option to cache config on disk
By default a copy of the configuration is writtent to a memory based
file while keepalived is loading/reloading. The new config_directory
global_defs option allows specifying a directory on whose filesystem
the config will be written instead.

This commit also adds support for systems that do not support
memfd_create() by using an anonymous file created on the /tmp
filesystem.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>