vrrp: Don't treat for IPv6 VIP non link local as error unless vrrp_strict set
While it is an error based on RFC5798 not to have the first IPv6 VIP being
link local, keepalived can run successfully even if the first VIP is not link
local. Since we have the vrrp_strict option, only treat it as a hard configuration
error if vrrp_strict is set, otherwise just log it.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>