Fix systemd notification when built without --enable-json
There was some optimisation of the size of the array used for function
pointers to signal handlers, which meant that the array wasn't large
enough if JSON was not enabled.

This commit removes that optimisation and makes the array large enough
to work with all signals. This will ensure that we don't have a similar
problem again when starting to use another new signal.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>