vrrp: allow vmac rename at reload
At reload, if we detect an existing vmac interface matches conf settings with
a different name and the name of the existing vmac interface is in the previous,
we delete existing to recreate a new vmac interface in order to rename.

Same behavior is kept at startup. If a vmac matching conf settings is already
found at startup, this vmac interface is used even the name is different.

Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>