config: keep the memfd copy of the config until reload or exit
One protection mechanism that keepalived uses is that if one of
the vrrp, checker or bfd processes exits, the parent process restarts
it. In order for that to work, the copy of the configuration must be
retained, so that it can be re-read.

Re-reading the original configuration files is not safe, because the
files might have been changed prior to a reload.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>