core: If epoll_wait() returns un unrecoverable error, exit and reload
github issue 1809 reported the healthchecker process getting error EINVAL
with a particular configuration; this looks as though it was memory corruption but
we have no way of tracking down how that happened. The process then sat in a loop
calling epoll_wait once a second, continually getting the same error.

This commit provides a way to escape the error if it happens again, by the process
terminating, and it will then be restarted by the parent process.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>