parser: Fix handling of configurations larger than 4096 bytes
Using the same file descriptor by the different keepalived processes
to read the memfd based configuration file doesn't work because the
file descriptors are not independant.

The child processes need to each open a separate file descriptor so
that the file can be read independantly by each process, rather than
reading in one affecting the file offset in the other processes.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>