build: Fix building with snmp on RedHat based s390x systems
While doing an rpmbuild worked, building keepalived outside an rpmbuild
environment was failing due to mismatched output of:
  net-snmp-config --base-cflags
and
  net-snmp-config --external-libs
caused by the RedHat patch net-snmp-ver-cflags.patch for building
net-snmp.

This commit detects the inconsistency, and modifies the output of
net-snmp-config --external-libs so that it is consistent with the
output of net-snmp-config --base-cflags.

See https://bugzilla.redhat.com/show_bug.cgi?id=1921988 for further
details.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>