Allow to override build date with SOURCE_DATE_EPOCH
in order to make builds reproducible when building from tarball.
See https://reproducible-builds.org/ for why this is good
and https://reproducible-builds.org/specs/source-date-epoch/
for the definition of this variable.

This date call only works with GNU date, as before.

This PR was done while working on reproducible builds for openSUSE.