core: Fix handling systemd notify
The handling of the child processes signalling the parent process that they
had completed reading the config file copy was not working reliably. The
problem was that all child processes were using the same signal number to
notify the parent, and so if more than one child process signalled the
parent before the parent had read the first signal, one of the signals was
lost.

As a temporary measure make the child processes use different signal numbers.
This may be changed in future to avoid the use of signals, such as using an
eventfd, but any change will be internal to keepalived.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>