core: Handle SELinux denying access to memfd files
If running in an SELinux controlled environment, permission needs
to be granted to a confined process to be able to create, read and
write memfd files.

This commit detects access being denied to a memfd file, and if so
uses an unnamed temporary file in the run filesystem instead.

In particular, the following SELinux permissions are needed:
   allow keepalived_t tmpfs_t:file { getattr open read write };

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>