doc: Add note to man page re moving VIP between VRRP instances
If a VIP (or virtual route or rule) is moved from one VRRP instance
to another when the configuration is reloaded, and both instances
are in master state, the VIP/route/rule can end up not being configured.

This commit adds a note to the keepalived(8) man page warning of this,
and suggests that two reloads should be done, the first to remove the VIP
and the second to add it to the other instance.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>