vrrp: Don't remove unweighted track_file/if/bfd from sync group members
Commit 3390697 - "vrrp: rewrote framework using list_head_t design"
cleared the track_file/if/bfd lists for any vrrp instance that was in a
sync group. This was due to the old list structure allocating memory
which had to be freed if the list was empty, but that is no longer
the case with list_head.

This commit is the equivalent of commit 459fc2b which was for track_scripts.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>