Travis-CI: Stop attempting to use it
So far as I can tell, Travis's offering of free credits is a hollow offer.

Having moved to travis-ci.com from travis-ci.org, we subsequently discovered
that, apart from the initial free 10,000 credits you either had to pay for
credits to use the service, or get an allocation of free OSS credits.

Following a request for free OSS credits 4 weeks ago, and receiving a response
about 3 hours later requesting certain information to confirm that keepalived
was an OSS project, and despite a couple of gentle reminders to ask how the
request was progressing, no response has been received.

Since no response has been forthcoming for 4 weeks, my only conclusions is that
free OSS credits are not being allocated. The project is therefore dropping
Travis-CI and has moved to Github actions, which is working very well.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>