vrrp: Fix changing vmac_xmit_base with IPv6 when reloading
If before a reload vmac_xmit_base is configured, then the VMAC interface
will not have an IPv6 link local address. If vmac_xmit_base is removed
when the keepalived configuration is reloaded, then keepalived needs to
add the relevant Ipv6 link local address.

Further, to be consistent, if vmac_xmit_base is added on a reload, then
remove the IPv6 link local address.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>