check: Rewrite genhash utility to be part of mainline daemon
Initial work on stream lining genhash code. There is no good reasons
for it to be a dedicated standalone tool. This extension create a
new keepalived command line option --genhash with all commands and
options from previous utility. Our goal is to use the same code path
in order to generate MD5SUM by submitting exactly the same code as
the one used for HTTP_GET and/or SSL_GET.