Add FREE_ARG_ON_RELOAD option to thread_add_read() and thread_add_write()
When sending an SMTP message, there is an smtp_t object allocated, and the
pointer to it is saved in the thread arg variable. If the thread is deleted
when keepalived reloads, the smtp_t object needs to be freed. Previously it
wasn't being freed resulting in a memory leak during reloads.

This commit adds a THREAD_FREE_ARG_ON_RELOAD flag to thread_add_read() and
thread_add_write(), so that when any such threads are freed, the smtp_t
object pointed to by the arg can be freed.

An alternative implementation would have been to add a cleanup function
pointer to thread_add_...() calls, and that could be an option in future.
This could handle both closing fds and freeing any allocated memory.

It is possible that in the future other thread_add_...() functions may also
need to support THREAD_FREE_ARG_ON_RELOAD.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>