all: Ensure reload doesn't overwrite config if reload in progress
The main process reads the configuration files and writes them to an
anonymous file which is then read by the child processes.

This commit ensures that the parent process doesn't start overwriting
the anonymous file before the child processes have completed reading it,
so that if two reloads are signalled in quick succession the child processes
don't read corrupted configurations.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>