core: fix detecting if build is in keepalived git tree
Alpine Linux appears to build keepalived in a sub-directory under the git
root of the Alpine Linux source tree. This means that
  git rev-parse --is-inside-work-tree
is true, and the but it is not keepalived's git tree, but rather the
Alpine Linux tree. `git describe --tags` then gives the git revision of
the Alpine tree, rather than keepalived's version, and as a consequence
the git version output by `keepalived -v` is the git version of Alpine
Linux instead of the git version of keepalived.

This commit checks that the git root is the root of the keepalived
source tree, and if not does not update lib/git-commit.h, thereby
preserving the keepalived git version.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>