ipvs: Stop using deprecated inet_aton()
It would have been simpler to use inet_pton() rather than getaddrinto(),
but that does not support IPv4 numbers-and-dots notation, whereas
inet_aton() does. We probably don't want to support numbers-and-dots
notation, but someone may be using it, and ipvsadm, since it uses
inet_aton(), does support numbers-and-dots notation for granularity mask.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>