Add data_use_instance global def
The filenames of the data dump files keepalived writes to after receiving
SIGUSR1 were always the same, and this caused problems if multiple
keepalived instances were running on a system.

This commit adds the data_use_instance global def which makes keepalived
include the instance name and network namespace name in the dump file names.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>