vrrp: Make VMAC IPv6 link local address mirror parent interface
If the link local address on the parent interface of a VMAC is
changed, keepalived will now change the link local address of the
VMAC so that it is the same as the parent interface's.

There was also an inconsistency in keepalived's behaviour. If the
parent interface's link local address was deleted, keepalived
generated a link local address for the VMAC based on the MAC address
of the parent interface. However, if when keepalived started the
parent interface didn't have a link local address, then one was not
assigned to the VMAC. keepalived will now generate a link local
address on startup if the parent interface does not have a link local
address.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>