vrrp: Check address label length < IFNAMSIZ
Issue #1951 identified that in certain circumstances address labels did
not work. While investigating that issue it was noticed that keepalived
did not notice address labels that were too long, and also if the label
was too long, it didn't terminate it with a NUL byte.

With this commit keepalvied now checks that the address label is not too
long, and it is is logs a config error and ignores the address. Previously
the address would not be added by the kernel due to the label being too
long, so the overall runtime functionality effectively remains the same.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>