vrrp: Add interface up/down debounce timers
There are  some users who experience very rapid interface up/down/up
state transitions, and would like the VRRP instance on the master not
to transition to backup state as soon as the interface goes down.

This commit adds interface down and up debounce timers so that a
master instance can remain in master state if its interface very
briefly transitions to down state and then reverts to up state.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>