vrrp: Drop ICMPV6 Router Solicitation messages from vmac interfaces
When we create a vmac interface, a short time afterwards the kernel
sends a router solicition message with the source MAC address of the
vmac interface. The problem is that this will upset snooping switches
if the VRRP instance is in backup state. Furthermore, we can't simply
move the packet onto the underlying interface since the ICMPV6
payload also contains the MAC address of the vmac interface. We can't
just change the MAC address in the ICMPV6 message, since there is
also a checksum which would need to be recalculated. The only
solution at the moment is to drop the packet. This shouldn't be a
problem since the underlying interface should have sent a Router
solicitation message when it came up.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>