vrrp: Don't lose some configuration faults
The following errors were being detected in vrrp_complete_instance() and
the VRRP instance was then supposed to be put into fault state since it
couldn't operate. However, the need to go to fault state was subsequently
being lost.

The configuration errors that were being lost were:
a) Configuring use of a VMAC on a non Ethernet interface
b) Attempting to use multicast on an interface that doesn't support it
c) Using an ipvlan without a source IP address
d) ipvlan address family not matching VRRP isntance's
e) VRID conflicts on an interface which could be deleted an recreated
     on a different interface
f) An interface specified for a VIP is the same as the VRRP instance's
     VMAC or another VRRP instance's VMAC.

This commit ensures that the VRRP instance will be put into, and remain
in, fault state, since it cannot successfully operate.

As can be seen from the list of circumstances above, they were very unlikely
to occur, but were possible.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>