vrrp: Bind IPv6 socket to multicast address
Previously IPv6 sockets were being bound to the ::1 address, since
trying to bind to the multicast address was failing.

The reason for failing has now been discovered to be that the scope_id
needed to be set (i.e. the interface index), since the multicast
addresses that we use are link-local multicast addresses.

This commit now sets the scope_id, so the socket can successfully be
bound to the multicast address.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>