vrrp: Is vrrp_t flags for checking if unicast is in use
Checking if the vrrp->unicast_peer list was not empty to check if
unicast was in use was entirely obvious what the code was doing.

This commit now checks the VRRP_FLAG_UNICAST in vrrp->flags to
check if unicast is in use, which should make what the code is
doing clearer.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>