all: Use sockaddr_t rather than struct sockaddr_storage
Struct sockaddr_storage is 128 bytes, whereas all keepalived needs
is a structure large enough for sockaddr_in and sockaddr_in6, which
means 28 bytes.

sockaddr_t is, by default, 28 bytes, but if --enable-sockaddr-storage
configure option is specifed, that sockaddr_t will be defined to be
sockaddr_storage, and so sockaddr_storage will be used.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>