install: Issue a warning if the DBus config file is unlikely to work
The standard DBus configuration expects the .conf files to be installed
in /etc/dbus-1/system.d. Unless --sysconfdir=/etc is specified at
configure time, they will be installed elsewhere, and keepalived DBus
is unlikely to work.

This commit makes 'make install' issue a warning if the .conf file is
installed in a non-standard location.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>