config: Fix location of default keepalived config file
Regardless of where keepalived was installed, it alsways used
/etc/keepalived/keepalived.conf as its default config file.

The default location for installing the config file is
/usr/local/share/etc/keepalived/keepalived.conf, although this
will be overridden if the --sysconfdir configuration option is
specified.

If keepalived now uses the default config file location, it will
attempt to open $(sysconfdir)/keepalived/keepalived.conf, and only
it that does not exist will it attempt to open
/etc/keepalived/keepalived.conf.

Since distros, when building their keepalived package specify
--sysconfdir=/etc, this commit will not make any difference to the
distros' builds.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>