vrrp dbus: Fix assert on shutdown after name lost
When DBus reports that the VRRP name has been lost on the session
bus, keepalived calls g_hash_table_foreach_remove(). However it called
the function again on shutdown causing an assert failure. This commit
ensures that g_hash_table_foreach_remove() is not called on shutdown
if it has already been called due to the name being lost.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>