rbtree: Fix walking Red Black trees and removing entries
It isn't safe to use the rb_for_each_entry loops if entries in the
tree are moved or removed, since the tree may be rebalanced.

If the entire tree is being deleted, then using
  rbtree_postorder_for_each_entry_safe
is OK, so long as rb_erase is not called for each entry, but the
entries are just FREE'd, and the root initialised at the end to
RB_BOOT. This is also more efficient than erasing each entry, since
there will be no rebalancing of the tree.

For removing multiple entries from the beginning of a cached tree,
rb_first_cached() can be called repeatedly with entries being erased,
until all the desired entries have been erased.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>