rbtree: Add rbtree_ka.h to minimise changes to Linux kernel files
The majority of the rbtree code is copied from the Linux kernel,
with a few changes made to them. To minimise the changes, making
it simpler to apply any updates to the code in the kernel, move
all keepalived specific code to rbtree_ka.h, and all keepalived
code should now include rbtree_ka.h rather than rbtree.h.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>