all: Add build time sanity changes of container_of parameters
container_of(ptr, type, member) is designed so that the type of
(type *)0->member is the same as the type of *ptr.

This commit adds checks to ensure that this is the case, as an
early mechanism of identifying potential bugs.

The code to make the check is copied from the Linux 5.15 kernel.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>