vrrp: always notify FIFO script of instance states at reload
Related to issue #2012.

Configuration reload restarts the notify FIFO script.

If the FIFO script has instructions to (un)set things (e.g. iptables
filter rules) depending on an instance state. And if unset things is
requested at script stop, then, at script restart, the new script is
not notified of the instance states. Things that needed to be set are not.

This patch always notifies FIFO script of instance states at reload.

Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>