vrrp: Add option to resend vrrp states on fifos after reload
Since keepalived restarts FIFOs scripts it is managing when a reload
occurs, it can be helpful to send the VRRP instance and group states
after a reload. This commit adds option fifo_write_vrrp_states_on_reload
to do that, and it means that what is written to the FIFOs with default
configuration does not change.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>